requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.8.1
certifi==2024.2.2
click==8.1.7
cloudinary==1.40.0
distro==1.9.0
Flask==3.0.3
gunicorn==20.0.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.25.1
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
